+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "python.pythonPath": "C:\\Users\\pm_\\AppData\\Local\\Programs\\Python\\Python38\\python.exe"
+}
