+++ cyclofit/templates/register02.html
@@ -17,34 +17,34 @@
         <form class="form-horizontal", role="form", autocomplete="off", method="POST", action="">
             {{ form.hidden_tag() }}
             <div class="form-group" style="margin-top: 34vh;">
-                {{ form.emergencyno.label(class='form-control-label') }}
-                {% if form.emergencyno.errors %}
-                    {{ form.emergencyno(class='form-control form-control-md is-invalid') }}
+                {{ form.area.label(class='form-control-label') }}
+                {% if form.area.errors %}
+                    {{ form.area(class='form-control form-control-md is-invalid') }}
                     <div class="invalid-feedback">
-                        {% for error in form.emergencyno.errors %}
+                        {% for error in form.area.errors %}
                             <span>
                                 {{ error }}
                             </span>
                         {% endfor %}
                     </div>
                 {% else %}
-                    {{ form.emergencyno(class='form-control form-control-md', placeholder='Enter Valid Number') }}
+                    {{ form.area(class='form-control form-control-md', placeholder='Area51') }}
                 {% endif %}
             </div>
             <div class="form-group">
-                {{ form.area.label(class='form-control-label') }}
+                {{ form.contactno.label(class='form-control-label') }}
                     
-                {% if form.area.errors %}
-                {{ form.area(class='form-control form-control-md is-invalid') }}
+                {% if form.contactno.errors %}
+                {{ form.contactno(class='form-control form-control-md is-invalid') }}
                 <div class="invalid-feedback">
-                    {% for error in form.area.errors %}
+                    {% for error in form.contactno.errors %}
                         <span>
                             {{ error }}
                         </span>
                     {% endfor %}
                 </div>
                 {% else %}
-                    {{ form.area(class='form-control form-control-md', placeholder='Area51') }}
+                    {{ form.contactno(class='form-control form-control-md', placeholder='Enter Valid Number') }}
                 {% endif %}
             </div>
             <div class="form-group">
@@ -64,19 +64,19 @@
                     {% endif %}
             </div>
             <div class="form-group">
-                {{ form.contactno.label(class='form-control-label') }}
+                {{ form.emergencyno.label(class='form-control-label') }}
                     
-                {% if form.contactno.errors %}
-                {{ form.contactno(class='mb-4 form-control form-control-md is-invalid') }}
+                {% if form.emergencyno.errors %}
+                {{ form.emergencyno(class='mb-4 form-control form-control-md is-invalid') }}
                 <div class="invalid-feedback">
-                    {% for error in form.contactno.errors %}
+                    {% for error in form.emergencyno.errors %}
                         <span>
                             {{ error }}
                         </span>
                     {% endfor %}
                 </div>
                 {% else %}
-                    {{ form.contactno(class='mb-4 form-control form-control-md', placeholder='Enter Valid Number') }}
+                    {{ form.emergencyno(class='mb-4 form-control form-control-md', placeholder='Enter Valid Number') }}
                 {% endif %}
             </div>
             <div class="col-12">
